+++ triplea/cli/analysis.py
@@ -25,23 +25,26 @@ from triplea.service.click_logger import logger
     multiple=True,
     required=True,
     help="""Generate graph and export it.
-                                The type of graph construction can be different. These include:
+The type of graph construction can be different. These include:
 
-                                store : It considers all the nodes and edges that are stored in the database
+store : It considers all the nodes and edges that are stored in the database
 
-                                gen-all : It considers all possible nodes and edges
+gen-all : It considers all possible nodes and edges
 
-                                article-topic : It considers article and topic as nodes and edges between them
+article-topic : It considers article and topic as nodes and edges between them
 
-                                article-author-affiliation : It considers article, author and affiliation as nodes and edges between them
+article-author-affiliation : It considers article, author and affiliation
+ as nodes and edges between them
 
-                                article-keyword : It considers article and keyword as nodes and edges between them
+article-keyword : It considers article and keyword as nodes and
+edges between them
 
-                                article-reference : It considers article and reference as nodes and edges between them
+article-reference : It considers article and reference as nodes and
+edges between them
 
-                                article-cited : It considers article and cited as nodes and edges between them
+article-cited : It considers article and cited as nodes and edges between them
 
-                                """,
+""",
 )
 @click.option(
     "--command",
@@ -98,7 +101,10 @@ def analysis(generate_type, command):
             l_edges.extend(graphdict["edges"])
 
         else:
-            logger.ERROR(f"Invalid value for '--generate' / '-g': {generate_type}")
+            logger.ERROR(
+                f"""Invalid value for
+                          '--generate' / '-g': {generate_type}"""
+            )
 
     print()
     # logger.DEBUG(f'Remove duplication in Nodes & Edges. ')