+++ triplea/cli/visualize.py
@@ -35,25 +35,33 @@ class Handler(http.server.SimpleHTTPRequestHandler):
     multiple=True,
     required=True,
     help="""Generate graph and export it.
-                                The type of graph construction can be different. These include:
+    The type of graph construction can be different. These include:
 
-                                store : It considers all the nodes and edges that are stored in the database
+    store :
+      It considers all the nodes and edges that are stored in the database
 
-                                gen-all : It considers all possible nodes and edges
+    gen-all :
+      It considers all possible nodes and edges
 
-                               article-topic : It considers article and topic as nodes and edges between them
+    article-topic :
+      It considers article and topic as nodes and edges between them
 
-                                article-author-affiliation : It considers article, author and affiliation as nodes and edges between them
+    article-author-affiliation :
+      It considers article, author and affiliation as nodes
+      and edges between them
 
-                                article-keyword : It considers article and keyword as nodes and edges between them
+    article-keyword :
+      It considers article and keyword as nodes and edges between them
 
-                                article-reference : It considers article and reference as nodes and edges between them
+    article-reference : It considers article and reference as nodes
+    and edges between them
 
-                                article-cited : It considers article and cited as nodes and edges between them
+    article-cited :
+      It considers article and cited as nodes and edges between them
 
-                                country-authorship :
+    country-authorship :
 
-                                """,
+    """,
 )
 @click.option("--port", "-p", "port", default=8000, help="port")
 def visualize(generate_type, port):
@@ -68,7 +76,9 @@ def visualize(generate_type, port):
             l_edges.extend(graphdict["edges"])
 
         elif g_type == "article-topic":
-            graphdict = gextract.graph_extractor(gextract.graph_extract_article_topic)
+            graphdict = gextract.graph_extractor(
+                gextract.graph_extract_article_topic
+                )
             l_nodes.extend(graphdict["nodes"])
             l_edges.extend(graphdict["edges"])
 
@@ -80,7 +90,9 @@ def visualize(generate_type, port):
             l_edges.extend(graphdict["edges"])
 
         elif g_type == "article-keyword":
-            graphdict = gextract.graph_extractor(gextract.graph_extract_article_keyword)
+            graphdict = gextract.graph_extractor(
+                gextract.graph_extract_article_keyword
+                )
             l_nodes.extend(graphdict["nodes"])
             l_edges.extend(graphdict["edges"])
 
@@ -92,16 +104,21 @@ def visualize(generate_type, port):
             l_edges.extend(graphdict["edges"])
 
         elif g_type == "article-cited":
-            graphdict = gextract.graph_extractor(gextract.graph_extract_article_cited)
+            graphdict = gextract.graph_extractor(
+                gextract.graph_extract_article_cited
+                )
             l_nodes.extend(graphdict["nodes"])
             l_edges.extend(graphdict["edges"])
 
         elif g_type == "country-authorship":
-            graphdict = gextract.graph_extractor(gextract.graph_extract_article_country)
+            graphdict = gextract.graph_extractor(
+                gextract.graph_extract_article_country
+                )
             l_nodes.extend(graphdict["nodes"])
             l_edges.extend(graphdict["edges"])
         else:
-            logger.ERROR(f"Invalid value for '--generate' / '-g': {generate_type}")
+            logger.ERROR(f"""Invalid value for
+                          '--generate' / '-g': {generate_type}""")
 
     # print()
     # logger.DEBUG(f'Remove duplication in Nodes & Edges. ')