+++ triplea/the_private_backyard3.py
@@ -2,6 +2,10 @@
 
 
 from tests.fixtures.graph_52 import graph52_instance
+from triplea.cli import export_graph
+
+import triplea.cli as cli
+
 from triplea.service.repository.import_file.triplea import import_triplea_json
 
 from triplea.service.graph.analysis.ganalysis import get_avg_shortest_path_length_per_node, get_clustering_coefficient_per_node
@@ -15,3 +19,5 @@ if __name__ == "__main__":
     
     result = get_avg_shortest_path_length_per_node(G)
     print(result[0])
+
+